A bit of a refactor around music
